Support parsing options out of normal order
Instead of requiring that all git-ssh wrapper options at the beginnning, this patch
allows for options to be specified anywhere in the git command. This allows for the following:

Prior, you would not be able to comfortably use shell completion, since once you specified a
wrapper option like --ssh, the completion script would fail to resolve other possible values.

$  git --ssh testing push -u origin testbranch

While that limitation still exists, it is out of the scope of this project. However, an alternative
can now be achieved by specifying the git-ssh options after all of the normal git options like so:

$  git push -u origin testbranch --ssh testing

This allows bash completion to work in the script all the way until the wrapper options are reached.

This also brings support for wrapper options interspersed between normal git options, though it is not
recommended and will not be fully supported moving forward.